before deleting supplier change its items supplier

+++ panelSupplierProcess.php
@@ -9,6 +9,16 @@ if (!isset($_POST['check'])) {
 
 if (!empty($_POST['box'])) {
     foreach ($_POST['box'] as $sid) {
+        $query = "select * from items where iSupplierId = '$sid'";
+        $result = mysqli_query($conn, $query) or die("Error in query: <mark>$query</mark> <p>". mysqli_error($conn));
+    
+        if (mysqli_num_rows($result) > 0) { // update supplier item records before delete
+            $supplier = mysqli_fetch_row(mysqli_query($conn, "select sId from suppliers"));
+
+            $query = "update items set iSupplierId = {$supplier[0]} where iSupplierId = '$sid'";
+            mysqli_query($conn, $query) or die("Error in query: <mark>$query</mark> <p>". mysqli_error($conn));
+        }
+
         $query = "delete from suppliers where sId = '$sid'";
         mysqli_query($conn, $query) or die("Error in query: <mark>$query</mark> <p>". mysqli_error($conn));
     }
